+++ src/api/objects/toot.ts
@@ -65,7 +65,7 @@ const PROPS_THAT_CHANGE: (keyof Toot)[] = [
 export interface SerializableToot extends mastodon.v1.Status {
     completedAt?: string;              // Timestamp a full deep inspection of the toot was completed
     followedTags?: MastodonTag[];      // Array of tags that the user follows that exist in this toot
-    isFollowed?: boolean;              // Whether the user follows the account that posted this toot
+    isFollowedAccount?: boolean;              // Whether the user follows the account that posted this toot
     reblog?: SerializableToot | null,  // The toot that was retooted (if any)
     reblogsBy?: mastodon.v1.Account[]; // The accounts that retooted this toot (if any)
     resolvedToot?: Toot;               // This Toot with URLs resolved to homeserver versions
@@ -134,7 +134,7 @@ export default class Toot implements TootObj {
     // extensions to mastodon.v1.Status. Most of these are set in completeProperties()
     completedAt?: string;
     followedTags?: mastodon.v1.Tag[];            // Array of tags that the user follows that exist in this toot
-    isFollowed?: boolean;                        // Whether the user follows the account that posted this toot
+    isFollowedAccount?: boolean;                        // Whether the user follows the account that posted this toot
     participatedTags?: TrendingTag[];            // Array of tags that the user has participated in that exist in this toot
     @Type(() => Account) reblogsBy!: Account[];  // The accounts that retooted this toot
     @Type(() => Toot) resolvedToot?: Toot;       // This Toot with URLs resolved to homeserver versions
@@ -184,7 +184,7 @@ export default class Toot implements TootObj {
         // Unique to fedialgo
         tootObj.completedAt = toot.completedAt;
         tootObj.followedTags = toot.followedTags;
-        tootObj.isFollowed = toot.isFollowed;
+        tootObj.isFollowedAccount = toot.isFollowedAccount;
         tootObj.reblog = toot.reblog ? Toot.build(toot.reblog) : undefined;
         tootObj.reblogsBy = (toot.reblogsBy ?? []).map(account => Account.build(account));
         tootObj.resolvedToot = toot.resolvedToot;
@@ -302,6 +302,11 @@ export default class Toot implements TootObj {
     // Return true if it's a direct message
     isDM(): boolean {
         return this.visibility === TootVisibility.DIRECT_MSG;
+    }
+
+    // Returns true if this toot is from a followed account or contains a followed tag
+    isFollowed(): boolean {
+        return !!(this.account.isFollowed || this.realToot().account.isFollowed || this.realToot().followedTags?.length)
     }
 
     // Return true if the toot has not been filtered out of the feed
@@ -484,8 +489,14 @@ export default class Toot implements TootObj {
         if (!this.shouldComplete()) return;
 
         this.muted ||= this.realAccount().webfingerURI in userData.mutedAccounts;
-        this.isFollowed ||= this.account.webfingerURI in userData.followedAccounts;
-        if (this.reblog) this.reblog.isFollowed ||= this.reblog.account.webfingerURI in userData.followedAccounts;
+        this.account.isFollowed ||= this.account.webfingerURI in userData.followedAccounts;
+        // TODO: get rid of isFollowedAccount and use account.isFollowed instead
+        this.isFollowedAccount ||= this.account.isFollowed;
+
+        if (this.reblog) {
+            this.reblog.account.isFollowed ||= this.reblog.account.webfingerURI in userData.followedAccounts;
+            this.reblog.isFollowedAccount = this.reblog.account.isFollowed;
+        }
 
         // Retoots never have their own tags, etc.
         const toot = this.realToot();
@@ -585,11 +596,12 @@ export default class Toot implements TootObj {
         // encounter the same Toot both in the user's feed as well as in a Trending toot list).
         Object.values(tootsByURI).forEach((uriToots) => {
             const firstCompleted = uriToots.find(toot => !!toot.completedAt);
-            const firstFollowedTags = uriToots.find(toot => !!toot.followedTags);
             const firstResolvedToot = uriToots.find(toot => !!toot.resolvedToot);
             const firstScoredToot = uriToots.find(toot => !!toot.scoreInfo);
             const firstTrendingLinks = uriToots.find(toot => !!toot.trendingLinks);
             const firstTrendingRankToot = uriToots.find(toot => !!toot.trendingRank);
+            // Account properties
+            const isAccountFollowed = uriToots.some(toot => toot.account.isFollowed)
             // Deal with tag arrays
             const allTrendingTags = uriToots.flatMap(toot => toot.trendingTags || []);
             const uniqueTrendingTags = uniquifyByProp(allTrendingTags, (tag) => tag.name);
@@ -610,7 +622,8 @@ export default class Toot implements TootObj {
                 toot.reblogsCount = propsThatChange.reblogsCount;
                 toot.repliesCount = propsThatChange.repliesCount;
                 // booleans can be ORed
-                toot.isFollowed = uriToots.some(toot => toot.isFollowed);
+                toot.isFollowedAccount = uriToots.some(toot => toot.isFollowedAccount);  // TODO: get rid of this
+                toot.account.isFollowed = isAccountFollowed;
                 toot.muted = uriToots.some(toot => toot.muted);
                 toot.sources = uniquify(uriToots.map(toot => toot.sources || []).flat());
                 // Set all toots to have all trending tags so when we uniquify we catch everything