+++ src/scorer/feed/diversity_feed_scorer.ts
@@ -28,7 +28,7 @@ export default class DiversityFeedScorer extends FeedScorer {
         // divided by the tag.numAccounts.
         sortedToots.forEach((toot) => {
             incrementCount(tootsPerAccount, toot.account.webfingerURI);
-            if (toot.reblog?.account) incrementCount(tootsPerAccount, toot.reblog.account.webfingerURI);
+            if (toot.reblog) incrementCount(tootsPerAccount, toot.reblog.account.webfingerURI);
 
             if (!toot.realToot().trendingTags) {
                 console.warn(`No trending tags for toot:`, toot);
@@ -70,7 +70,7 @@ export default class DiversityFeedScorer extends FeedScorer {
                     decrementCount(trendingTagPenalty, tag.name, trendingTagIncrement[tag.name]);
                     const logStr = `penalty: -${trendingTagPenalty[tag.name]}, increment: ${trendingTagIncrement[tag.name]}, scored so far: ${tootsWithTagScoredSoFar[tag.name]} for toot ${toot.realToot().describe()}`;
 
-                    if (toot.isFollowed || toot.reblog?.isFollowed) {
+                    if (toot.account.isFollowed || toot.reblog?.account.isFollowed) {
                         // if (toot.trendingTags?.length) traceLog(`${this.logPrefix()} Not penalizing followed toot:`, toot.realToot().describe());
                     } else if (tootsWithTagScoredSoFar[tag.name] > Config.minTrendingTagTootsForPenalty) {
                         // ...but only apply the penalty after MIN_TRENDING_TAGS_FOR_PENALTY toots have been passed over